Split authentication DTOs into their own files and move to dto/ folder
