+++ src/components/authentication/dto/index.ts
@@ -1,1 +1,4 @@
+export * from './login.dto';
+export * from './passwords.dto';
 export * from './register.dto';
+export * from './session.dto';
