+++ src/components/authorization/dto/index.ts
@@ -1,1 +1,2 @@
 export * from './role.dto';
+export * from './powers';
